refactor(calculator): extract derivative computation into derivatives.py
